# Backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserBody": {
				"properties": {
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUser": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LeaderboardEntryDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"rank": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"username",
					"score",
					"rank",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Color": {
				"enum": [
					"BLACK",
					"WHITE"
				],
				"type": "string"
			},
			"GameDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"owner": {
						"$ref": "#/components/schemas/UserDTO"
					},
					"public": {
						"type": "boolean"
					},
					"owner_win": {
						"type": "number",
						"format": "double"
					},
					"creation_date": {
						"type": "number",
						"format": "double"
					},
					"date_end": {
						"type": "number",
						"format": "double"
					},
					"owner_color": {
						"$ref": "#/components/schemas/Color"
					}
				},
				"required": [
					"id",
					"owner",
					"public",
					"owner_win",
					"creation_date",
					"owner_color"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StatsDTO": {
				"properties": {
					"totalGames": {
						"type": "number",
						"format": "double"
					},
					"wins": {
						"type": "number",
						"format": "double"
					},
					"winRate": {
						"type": "number",
						"format": "double"
					},
					"publicGames": {
						"type": "number",
						"format": "double"
					},
					"privateGames": {
						"type": "number",
						"format": "double"
					},
					"gamesAsWhite": {
						"type": "number",
						"format": "double"
					},
					"gamesAsBlack": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"totalGames",
					"wins",
					"winRate",
					"publicGames",
					"privateGames",
					"gamesAsWhite",
					"gamesAsBlack"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PieceType": {
				"enum": [
					"PAWN",
					"ROOK",
					"KNIGHT",
					"BISHOP",
					"QUEEN",
					"KING"
				],
				"type": "string"
			},
			"Piece": {
				"properties": {
					"pieceType": {
						"$ref": "#/components/schemas/PieceType"
					},
					"color": {
						"$ref": "#/components/schemas/Color"
					},
					"i": {
						"type": "number",
						"format": "double"
					},
					"j": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pieceType",
					"color",
					"i",
					"j"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Case": {
				"properties": {
					"color": {
						"$ref": "#/components/schemas/Color"
					},
					"piece": {
						"$ref": "#/components/schemas/Piece"
					}
				},
				"required": [
					"color"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReturnAction": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"result": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"listCase": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Case"
							},
							"type": "array"
						},
						"type": "array"
					},
					"turn": {
						"$ref": "#/components/schemas/Color"
					},
					"pieceKilled": {
						"items": {
							"$ref": "#/components/schemas/Piece"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"result",
					"listCase",
					"turn",
					"pieceKilled"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MovePiece": {
				"properties": {
					"i": {
						"type": "number",
						"format": "double"
					},
					"j": {
						"type": "number",
						"format": "double"
					},
					"toI": {
						"type": "number",
						"format": "double"
					},
					"toJ": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"i",
					"j",
					"toI",
					"toJ"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReturnGameAction": {
				"properties": {
					"i": {
						"type": "number",
						"format": "double"
					},
					"j": {
						"type": "number",
						"format": "double"
					},
					"toI": {
						"type": "number",
						"format": "double"
					},
					"toJ": {
						"type": "number",
						"format": "double"
					},
					"piece": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"pieceKilled": {
						"items": {
							"$ref": "#/components/schemas/Piece"
						},
						"type": "array"
					},
					"step": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"i",
					"j",
					"toI",
					"toJ",
					"pieceKilled",
					"step"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChessReplay": {
				"properties": {
					"ownerUsername": {
						"type": "string"
					},
					"ownerColor": {
						"$ref": "#/components/schemas/Color"
					},
					"actions": {
						"items": {
							"$ref": "#/components/schemas/ReturnGameAction"
						},
						"type": "array"
					}
				},
				"required": [
					"ownerUsername",
					"ownerColor",
					"actions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateGameBody": {
				"properties": {
					"isPublic": {
						"type": "boolean"
					},
					"ownerColor": {
						"$ref": "#/components/schemas/Color"
					}
				},
				"required": [
					"isPublic",
					"ownerColor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FormatedGame": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"listCase": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Case"
							},
							"type": "array"
						},
						"type": "array"
					},
					"turn": {
						"$ref": "#/components/schemas/Color"
					},
					"pieceKilled": {
						"items": {
							"$ref": "#/components/schemas/Piece"
						},
						"type": "array"
					},
					"result": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"ownerColor": {
						"$ref": "#/components/schemas/Color"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ChessLocation": {
				"properties": {
					"i": {
						"type": "number",
						"format": "double"
					},
					"j": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"i",
					"j"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthBody": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"grant_type": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password",
					"grant_type"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserBody"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:write"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUser"
							}
						}
					}
				}
			}
		},
		"/leaderboards": {
			"get": {
				"operationId": "GetLeaderboard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LeaderboardEntryDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [],
				"parameters": []
			}
		},
		"/leaderboards/history/me": {
			"get": {
				"operationId": "GetUserGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/leaderboards/history/{userId}": {
			"get": {
				"operationId": "GetLeaderboardUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/leaderboards/stats/me": {
			"get": {
				"operationId": "GetMyStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StatsDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/leaderboards/stats/{userId}": {
			"get": {
				"operationId": "GetUserStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StatsDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games/piece/move": {
			"post": {
				"operationId": "MovePiece",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReturnAction"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MovePiece"
							}
						}
					}
				}
			}
		},
		"/games/replay/{id}": {
			"get": {
				"operationId": "GetReplay",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChessReplay"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games/create": {
			"post": {
				"operationId": "CreateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"gameId": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"gameId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateGameBody"
							}
						}
					}
				}
			}
		},
		"/games": {
			"get": {
				"operationId": "GetGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FormatedGame"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/games/current-game": {
			"get": {
				"operationId": "GetUserGameId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"gameId": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"gameId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/games/piece/possible-move": {
			"post": {
				"operationId": "GetPossibleMove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ChessLocation"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"j": {
										"type": "number",
										"format": "double"
									},
									"i": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"j",
									"i"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/games/piece/upgrade": {
			"post": {
				"operationId": "UpgradePiece",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FormatedGame"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"piece": {
										"$ref": "#/components/schemas/PieceType"
									}
								},
								"required": [
									"piece"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Auth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}